+++ source/components/Handlers.tsx
@@ -5,6 +5,7 @@ import {Text} from 'ink';
 import options from '../data/options.js';
 import VideoDownloader from './video_downloader/VideoDownloader.js';
 import Mp3Downloader from './mp3_downloader/Mp3Downloader.js';
+import BackgroundRemover from './background_remover/BackgroundRemover.js';
 type OptionItem = {
 	label: string;
 	value: string;
@@ -28,7 +29,7 @@ function Displayer({
 		case 'mp3converter':
 			return <Mp3Downloader setChoice={setChoice} />;
 		case 'bg_image':
-			return <Text>You chose to download audio.</Text>;
+			return <BackgroundRemover setChoice={setChoice} />;
 		case 'img_pdf':
 			return <Text>You chose to download an image.</Text>;
 		case 'pdf_fusion':