+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refractoring to migrate to Ink 
 - Migrate from JS to Typescript
 - Add Youtube Video to MP3 Converter
+- Add Background Image remover
 
 ## [1.0.0] - 2024-03-04
 ### Added